pyproject.toml
[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
name = "root"
version = "0.0.0"
description="flask template"
[tool.poetry.dependencies]
flask = "==1.0.2"
python = "^3.8"
werkzeug = "^1.0.1"
psycopg2-binary = "^2.9.3"
